# Optimize Exceptions.throwIfFatal()