`Exceptions.throwIfFatal()` used in almost all RxJava operators.

Basically, in this PR I've just divided `throwable` into 3 groups: `RuntimeException`, `Error` and `Exception` (we don't need to check them).

With this PR `Exceptions.throwIfFatal()` becomes ~7% faster than current implementation in case of non-fatal exceptions (because if it's fatal — app will be crashed :smiley_cat:) on the Oracle JRE 1.8.0_45.

I guess, on Android it will be even more faster! (I am not sure in quality of our JITs/VMs).

I've checked the bytecode after compilation in hope that javac will optimize `Exceptions.throwIfFatal()` into same construction, but it does not perform any optimizations over this method at compile time, so the only hope is JIT and even after JIT, optimized version slightly faster (~7%).

Results of the benchmarks:

```
OPTIMIZED VERSION:

Benchmark                                                               Mode   Samples        Score  Score error    Units
r.e.ExceptionsThrowIfFatalPerf.throwIfFatalNonFatalCheckedException     avgt         5        2.569        0.389    ns/op
r.e.ExceptionsThrowIfFatalPerf.throwIfFatalNonFatalError                avgt         5        2.332        0.030    ns/op
r.e.ExceptionsThrowIfFatalPerf.throwIfFatalNonFatalRuntimeException     avgt         5        2.511        0.037    ns/op

Benchmark                                                               Mode   Samples        Score  Score error    Units
r.e.ExceptionsThrowIfFatalPerf.throwIfFatalNonFatalCheckedException    thrpt         5 395071210.861  1644491.739    ops/s
r.e.ExceptionsThrowIfFatalPerf.throwIfFatalNonFatalError               thrpt         5 429656328.864  3230560.292    ops/s
r.e.ExceptionsThrowIfFatalPerf.throwIfFatalNonFatalRuntimeException    thrpt         5 398472020.961  3961447.943    ops/s

OLD VERSION:

Benchmark                                                               Mode   Samples        Score  Score error    Units
r.e.ExceptionsThrowIfFatalPerf.throwIfFatalNonFatalCheckedException     avgt         5        2.812        0.954    ns/op
r.e.ExceptionsThrowIfFatalPerf.throwIfFatalNonFatalError                avgt         5        2.338        0.019    ns/op
r.e.ExceptionsThrowIfFatalPerf.throwIfFatalNonFatalRuntimeException     avgt         5        2.771        0.064    ns/op

Benchmark                                                               Mode   Samples        Score  Score error    Units
r.e.ExceptionsThrowIfFatalPerf.throwIfFatalNonFatalCheckedException    thrpt         5 370211340.236  9093081.248    ops/s
r.e.ExceptionsThrowIfFatalPerf.throwIfFatalNonFatalError               thrpt         5 427943417.424  6168313.153    ops/s
r.e.ExceptionsThrowIfFatalPerf.throwIfFatalNonFatalRuntimeException    thrpt         5 368255154.316  4520997.005    ops/s
```