@davidmoten 2015-09-25
The path for `OnErrorNotImplementedException` or `OnErrorFailedException` will be slower because they are checked for being a `RuntimeException` first. If it's a `RuntimeException` and not one of those then it will be faster because checks against `Error` types will not occur. I think this contribution is probably ok but some estimate of the most commonly used paths through this code should probably be made. In my opinion the path taken by an `Error` will not in general be taken anywhere near as frequently as the path by a `RuntimeException` so the optimization should favour the `RuntimeException` path. The question remains then what `RuntimeException`s are most likely to pass through? I'm assuming `Exception`s won't be travelling by this path very often because I can't think where in our operators we might be throwing checked exceptions (someone can enlighten me of course).