@akarnokd 2015-09-25
Your benchmark results are within the error margin to each other so I'd not consider this as a clear-cut improvement. Throwing exceptions is so expensive that a simple `instanceof` check is overshadowed. And besides, optimizing for the exception path is questionable because you are not delivering the exceptions as part of the main business logic.