@stevegury 2015-09-25
I agree with @akarnokd here, the gain is not clear enough to justify the logic rewrite.