@artem-zinnatullin 2015-09-26
Okay, I agree with all points.

Checked some amount of libraries with RxJava support. Only small amount of them use `Exceptions.throwIfFatal()` in their implementations of `OnSubscribe`, so the case of passing checked `Exception` into the `throwIfFatal()` is pretty rare.

Also, as said in comments, exceptions are not / should not frequent cases of the application lifecycle, so the performance of this method does not really matter.

Thanks for the comments!